Restart Button + Game Over
Made a restart button for the pause menu as well as restart and Quit buttons for the game over screen.
However, the issue of the player being immortal is still present. (I thought this was something you were going to fix? If you can't it's all good, it's just that I don't know how to take back the gift of eternal life from the player character).